requirements.txt
requests==2.31.0
telnyx==2.0.0
python-dotenv==1.0.0
six==1.16.0
Flask-Limiter==3.3.1
uvicorn==0.23.2
fastapi==0.100.1
pydantic==2.1.1
slowapi==0.1.8
werkzeug==2.3.6
fastapi_limiter==0.1.5